Add files via upload
These are Django code. The first 'urls.py' is under the file 'my-project'. The second 'urls.py' and 'views.py' is under the file 'posts'. I use the 'my-project'-'urls.py' to include my new url and approach 'views.py' from the new url file (second 'urls.py'). It seems like that I use the wrong way to 'import' them. I change 'import . from views' to 'import views from *'. However, it does not work and it says that 'views' is not a name.